namespace under GraphQL, minor cleanups
