+++ bin/console
@@ -1,14 +1,12 @@
 #!/usr/bin/env ruby
 
 require "bundler/setup"
-require "graphql_metrics"
+require 'graphql'
+require "graphql/metrics"
 
 # You can add fixtures and/or initialization code here to make experimenting
 # with your gem easier. You can also use a different console, if you like.
 
 # (If you use this, don't forget to add pry to your Gemfile!)
-# require "pry"
-# Pry.start
-
-require "irb"
-IRB.start(__FILE__)
+require "pry"
+Pry.start
