+++ lib/graphql/metrics/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module GraphQL
+  module Metrics
+    VERSION = "3.0.0"
+  end
+end
