--- lib/graphql_metrics/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module GraphQLMetrics
-  VERSION = "3.0.0"
-end
